Fix on delete asset and rename the helper_functions to utils
