+++ assets/list_assets.py
@@ -3,7 +3,7 @@ import sublime
 import mdpopups
 import re
 
-from .helper_functions import tabulate_from_markdown,AssetSnapshot
+from .utils import tabulate_from_markdown,AssetSnapshot
 from pieces_os_client import *
 from pieces.settings import PiecesSettings
